Fix PrayerView
